ruleguard/typematch: handle builtin types in opNamed correctly
"error" type is *types.Named, but it doesn't have the Package.
This can lead to a nil pointer dereference if we don't handle that case.

Since we never emit opNamed for builtin types, we can fix the
problem by doing an early "return false" for types without a package.

Fixes #61

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>