ruleguard,dsl: add experimental support for const value filters
Added Value expression field that can be used to check the submatch value.
Since `go/types` has some const-folding out-of-the-box, we get it too.

Right now only `int` value types are supported, but we could extend the
set of types later.

This is a proof-of-concept that can be useful in #96.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>