ruleguard: fix interpolation of vars with common prefix
If pattern had variables where one of them is a prefix of another,
there was a chance of incorrect results.

This change makes renderMessage() sort keys and interpolate longer
var names first, so we avoid that risk.

Scratch slice is used to do the sorting, so no extra allocations
is added. As successfull matching is far less frequent than
a failed (or rejected) match, sorting should not affect the performance.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>