ruleguard,dsl: add Var.Node filters
Add the ability to assert the submatch node AST type.

A new `Var.Node.Is()` function accepts the type name string,
like `Ident` (ast.Ident), `Expr` (ast.Expr) and so on.
For any unexpected arguments the rule compilation will fail.

We may add more node categories that are not declared in
`go/ast` package for our convenience.

Rationale: sometimes we need to check not only the types,
but also the "shape" of the match. For example, we may want
to enforce some func arguments to be literal strings.
Right now we can only do this by doing a regexp check
that may look like `^["`].*["`]` but that's not reliable, nor efficient.
What we can do is to add a negated `*ast.BasicLit` constraint,
optionally with a type check if func signature accepts `interface{}`.

	// Match any non-literal $s
	Where(!m["s"].Node.Is(`BasicLit`))

Note that `m["s"].Const` is not identical here. It would accept
const expressions like `"a" + "b"` as well.

Not to be confused with an actual type. It's just a
syntactical check whether a matched node belongs to some
syntax category as defined by the `go/ast` package.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>