analyzer: add -enable and -disable flags
`-enable` and `-disable` are a comma separated lists.
It's not very convenient to specify a long list in bash,
but since most configuration is done in yaml (golangci-lint),
it should be somewhat readable:

	disable: >-
		A,
		B,
		C

Another option:

	disable: "\
		A,\
		B,\
		C"

Useful in combination with ruleguard modules.
Rules from modules can be disabled via -disable flag.

Both `-enable` and `-disable` flags can be debugged with
`-debug-enable-disable `flag (prints whether a group was enabled/disabled).

Refs #149