ruleguard: add support for structured doc comments
A matcher function can now use 4 special pragmas:

	* `doc:summary` for a short description
	* `doc:before` for an example of a "bad code"
	* `doc:after` for a fixed code example
	* `doc:tags` for a list of custom tags

It's planned to make it possible for the integrating app
to access that information in a convenient way.

This PR is moslty about enabling on a parsing level.

These 4 fields (summary, before, after, tags) are selected
as something that is close to the `go-critic` checker documentation model.