internal/gogrep: allow `f($*_)` to match variadic calls
We rename `CallExpr` to `NonVariadicCallExpr` -- it rejects variadic calls.
We keep `VariadicCallExpr` as is.
We introduce a new `CallExpr` op that matches both normal and variadic calls.

`CallExpr` is generated for call patterns that should accept both kinds of calls.
So if the last pattern argument is seq match var, we emit `CallExpr`.

Fixes #237