ruleguard: associate rule line with a pattern string
In a multi-pattern rule, it's more useful to give a source line
of a respective pattern.

Consider this example:

         //line1: m.Match(
         //line2:        `foo()`,
         //line3:        `bar()`,
         //line4: )

Previously, if either `foo()` or `bar()` would match, we would
have a matchInfo rule line set to `line1`, since it's a
rule definition line (`Match` expression location).

Now we'll have a rule line set to `line1` if `foo()` matches,
otherwise `line2` if `bar()` matches. This is useful during
both debugging and rules usage.