analyzer/testdata: use `\Q` regexp quoting everywhere
In analysis test we're using `want` directive to match warnings.
I personally thing that 99% of time we need a string matching,
not a regexp matching. So this test checks that we do `\Q`
escaping everywhere, making our patterns matching string "as is";
so you don't have to escape `(` and `)` in the patterns.