dsl: add ExprType.HasMethod()
HasMethod reports whether a type has a given method.
Unlike Implements(), it will work for both value and pointer types.

fn argument is a function signature, like `WriteString(string) (int, error)`.
It can also be in form of a method reference for importable types: `io.StringWriter.WriteString`.

To avoid confusion with Implements() method, here is a hint when to use which:

	- To check if it's possible to call F on x, use HasMethod(F)
	- To check if x can be passed as I interface, use Implements(I)

Refs #331