ruleguard: autofix some interpolated expressions
Given this pattern:

	io.WriteString($w, $s)

And this input string:

	io.WriteString(&buf, s)

With this replacement template:

	$w.WriteString($s)

We'll get this result:

	&buf.WriteString(s)

That's incorrect.

To avoid that, ruleguard attempts to fix some interpolated
expressions when it can understand the context.

In the cases above, it sees that variable is followed by `.`
while `$w` contains unary `&` expression. It knows that
`$w.x` is identical to `(&$w).w`, so we can simply insert
`buf` instead of `&buf` and get `buf.WriteString(s)`.

This greatly simplifies the creation of reliable rules
that just work.