analyzer: use sync.Pool to re-use ruleguard runner state
For some test runs, this allows to allocate ~10 state objects
instead of ~3000 without such pool.

Allocating a runner state is an expensive operation, we
want to re-use as much state as possible to make ruleguard faster.

This patch should be used as an example of how ruleguard
can be integrated into `go/analysis` framework while still
having a reusable state between the engine runs.